Update Patch kernel power to 2.5 to avoid numerical errors.
For large exponents and images above 2^10 pixels per dimension, the
kernel values at pixels which are distant from the center are very low
(~1e-10), yielding large numerical errors when dividing chi and erf.
The current solution is to lower the exponent a bit, which makes a less
smooth boundary but avoids these artifacts.

Another option is to make the exponent vary by the resolution, such that
the lowest value of the kernel will be M, where M is around 1e-7, 1e-8.
This will create the smoothest results but may include varying results
when working with different working sizes.