Fix invalid test with LTMMTexture and half_float.
Using image.setTo with a scalar treated the number as an unsigned short
instead of half_float. This doesn't work on a device since as an half
float it is a negative number which clamps to zero when reading the data
back from the GPU. To mitigate, used std::transform to fill each cell
explicitly.