Modify -[LTImageProcessor process] to return void.
This effectively frees the LTImageProcessor implementor from returning
the output in the -process method, and removes all the classes that
implement LTImageProcessorOutput protocol and the protocol itself. The
idea behind creating this return value was that processors that need to
create an output will return them via a return value of the -process
method (such as crop). The new design is to make that void, so that such
processors will accept an (LTTexture **) output type and therefore will
be able to return a newly created texture (if needed) without using a
return value for -process. Hopefully, this will make a clearer
interface, where previously users didn't know which texture to look at
after processing - the one returned from -process or the one they
initialized the processor with as an output(s).