LTBidirectionalMap: fix nullified object bug.
The scenario was adding a key-value pair to the map, where the key is
not held by an external source. Since NSDictionary copies the key and
the NSMapTable held a weak reference to it, it was destroyed in the next
drain of the autorelease pool.

The fix is to hold the values in the NSMapTable strongly. This should
not cause any memory management trouble such as retain cycles.