LTMMTexture: use correct CVPixelBuffer locking flags.
Until now, both read and write lock used the zero flag, which may
invalidate internal caches. The reason was that, AFAIR, the read only
flag caused the GPU data to not appear on the CPU side. However, this
now seems to work as expected.

Ran relevant unit tests on device(s) as well to validate.