NSError+LTKit: migrate to more generic factory methods.
The rationale behind this commit is to avoid over-specification when
creating errors for specific error codes to allow more flexibility.

For example, the factory method
+[lt_fileRemovalFailedErrorWithPath:underlyingError:] assumes that only
path and underlyingError are arguments that can be supplied with
NSErrorCodeFileNotFound, but what happens if one wants to attach a
description or multiple underlying errors?

The changes in this commit amends that by creating more generic NSError
factory method that accept various userInfo arguments but also the error
code to create the error from.

Additionally, the keys of the userInfo dictionary are now based more on
NSFoundation's keys than on our own custom ones.