LTBoundaryCondition: correct behavior near edges.
Previously, the boundary condition returned a float value for sampling
but handled the signal as an integral grid. This caused problems when
trying to sample the rightmost sample where only a singular value
returned it, while any other epsilon deviation from it yielded the value
left to that sample.

The new approach handles the signal as a continuous grid, so the values
returned can be in the range [0, N] (so extra care is required when
this method returns N when sampling from arrays to avoid out-of-bounds).
This approach is more symmetrical for sampling points near the edge of
the signal.