LTQuadMixerProcessor: Allow usage of last frag data.
This commit enables the LTQuadMixerProcessor to use the last frag data
for mixing rather than accessing a specific texture. This way it is
possible to perform several consecutive render passes without losing the
result of previous render passes.

In particular, this commit is required in order to improve the
performance of the ENMemeProcessor.