NSError+Factory: make NSError+LTKit project-agnostic.
The purpose of this commit is to split NSError+LTKit to two different
files: NSError+Factory, which is used to easily create NSError objects
with various properties, and NSErrorCodes+LTKit, which contains the
error codes that are populated and used in LTKit and throughout LT
products in general.

Additional instructions on how to add and register error codes to new
projects are contained in NSErrorCodes+LTKit.