LTRef: add lt::Ref instead of specifying default deleters.
As it turns out, its quite cumbersome to template unique_ptr on the
element type and not the pointer type. Therefore, a new lt::Ref class is
created which is templated on the pointer type contains a unique_ptr
that holds the reference. This also allows an easier way to pass
references to other functions by implementing an implicit cast operator
to the underlying reference.