LTGPUQueue: avoid deadlock by not waiting on dispatch group.
In some scenarios it is possible for dealloc to be called on a thread
that belongs to the dispatch group. This will cause a deadlock.

Since we do not use weak references to self, any block that is
dispatched to the queues will hold self strongly, so if the deallocation
runs on one of the queues of the GPU queue, it must be called from the
last dispatch on that queue, and once it will finish execution the queue
will not receive additional requests.

Additionally, even if there were a block that would run after dealloc,
there's no need to block the deallocation of the GPU queue, since GCD's
documentation on dispatch queues say:
"When your application no longer needs the dispatch queue, it should
release it with the dispatch_release function. Any pending blocks
submitted to a queue hold a reference to that queue, so the queue is not
deallocated until all pending blocks have completed."

So since the dispatch queue is refcounted, the remaining blocks will
execute even after the queue is deallocated.