LTGLPixelFormat: initial commit.
This enumeration will replace the triplet (format, precision, internal
format) with a single enumeration that provides a well-defined OpenGL
internal storage format, as well as methods to access the required
format, precision and internal format OpenGL enumerations. Additionally,
it provides interoperability formats for other libraries such as OpenCV
and CoreVideo.