LTParameterizedObjectFactory: Return LTCompoundParameterizedObject.
This commit updates the LTParameterizedObjectFactory to return
LTCompoundParameterizedObject objects rather than objects implementing
the LTParameterizedObject protocol. This is required in order to enable
objects using the factory to create new LTCompoundParameterizedObjects
from existing ones, with modified mappings. Without this commit, no
assumption on the actual class may be made. In the course of this change, the
LTParameterizedObjectFactory is renamed to LTCompoundParameterizedObjectFactory.