LTTexture: simplify initialization and deallocation pattern.
LTTexture used to have a virtual constructor pattern: its designated
initializer calls an abstract create method, which is overridden by
child classes to perform the actual resource creation.

The problem with this approach (among other things) is the great
difficulty of implementing derived classes that need additional
initialization parameters and/or initialization flows that are different
from the hard-coded base implementation.

This commit eliminates the virtual create and destroy methods, and
instead introduces a designated initializer. It is restricted to
initialization of properties declared by the base class itself, and
nothing else. Derived classes must call it, and perform the actual
resource allocation, by implementing an override, and optionally by
wrapping it in other implementation specific initializers.

For backward compatibility, other "helper" initializers are left as they
are in the base implementation.