LTIterativeProcessingStrategy: properly reset iteration count.
The variable responsible for the iterations count - currentIteration was
not set to zero when reseting the strategy state. This means that
running the strategy more than once caused the iterations count to start
at the total number of iterations as it was set on the previous run (if
iterationsPerOutput was not changed between two following runs, this
means that the second run won't be executed at all).