LTViewDelegates: Add method for requesting stop of touch handling.
This commit adds the ltViewStopTouchHandling method to the
LTViewTouchDelegate protocol. The method is currently called by the
LTView when the internal LTEAGLView is recreated due to calls to
layoutSubviews in order to compensate for the lack of calls to the
touchesEnded:withEvent: or touchesCancelled:withEvent: methods of LTView
(functioning as UIResponder) in this scenario.