LTContentInteraction: Initial commit.
This commit adds the LTContentInteraction protocols. In particular,
LTInteractionModeProvider, a protocol to be implemented by objects
providing the mode defining the user interaction using gestures and/or
content touch events, LTInteractionModeDelegate, a protocol to be
implemented by objects maintaining the mode defining the user
interaction via gestures and/or touches on a content view, and
LTContentInteractionManager, a protocol to be implemented by objects
managing the user interaction via gestures and/or touches on a content
view.