LTViewPixelGrid: Rename to LTGridDrawingManager.
This commit renames the LTViewPixelGrid class to LTGridDrawingManager.
This renaming is performed since a) the class is agnostic to (the
former) LTView (now LTPresentationView), b) the class is agnostic to the
notion of pixels in terms of the content presented by the
LTPresentationView, and c) the name "Grid" is misleading, suggesting the
object is a value class rather than a mutable class managing the grid
drawing.