DVNRenderStageConfiguration: Initial commit.
This commit DVNRenderStageConfiguration, a value class representing the
configuration of the render stage of the DVNPipeline. The configuration
consists of the source code of the vertex and fragment shaders used by
the pipeline, as well as a mapping of shader sampler uniforms to
textures and the values of the shader uniform variables to use.