LTPainter: prevent leak in tests.
The leak is triggered after OCMock is updated to OCMock 3.3.1.
Mock of painter's delegate is invoked with painter (among other things)
as delegate's argument, and the argument is strongly help for later
verification by OCMock, thus preventing timely deallocation of the
painter, along with other OpenGL resources.

The solution is to kill the delegate in afterEach.

The touchCollector change is harmless, and is added just to be on the
safe side in the future.