LTParameterizedObjectConstructor: Fix flaw in optimization.
This commit fixes a flaw in the optimization of
LTParameterizedObjectConstructor. Before this commit, the
LTMutableEuclideanSpline object attached to the (only internally known)
subclass of LTControlPointModel could have been updated by two different
LTParameterizedObjectConstructor objects, in case both of them were
initialized with aforementioned model. After this commit, it is
guaranteed that the spline cannot be modified anymore by
any LTParameterizedObjectConstructor object or any object in general
(assuming no one uses reflection to find out that the returned
id<LTParameterizedObject> object is actually a LTMutableEuclideanSpline.