BZRValidatricksReceiptValidator: Fix server URL and validator endpoint.
The validation request is sent to a URL composed of the validation
server URL and a validation endpoint. The two are composed by the HTTP
client using NSURL composition methods. NSURL composition method expects
the first parameter to be a URL ending with a '/' character, otherwise
it discards the URL's last path component. The second parameter is
expected to be a path relative to the first parameter, but if that path
starts with a '/' it treats it as an absolute path and discards the
entire path of the first parameter URL.