Bazaar: Fix validatricks date time to NSDate transformation.
The fix is required since NSTimeInterval is assumed to be in seconds
from 1/01/1970 00:00:00 and the value provided by Validatricks server is
in milliseconds from the same point in time. The transformer assumed it
to also be in seconds. The bug is fixed by adding a new transformer to
`NSValueTransformer+Validatricks` that correctly converts Validatricks
date time values to NSDate values and use it instead of the previously
transformer used.