Foundations: Create Foundations workspace and FoundationsTestsHost app.
The Foundations workspace is used as a container for Foundations sub
projects. The main reason to use it is to provide a unified tests host
application. A unified tests host has two main benefits:
1) Sub projects that require a standalone tests host application don't
need to create their own host they can now just use FoundationTestsHost.
2) The unified tests host can be used to run tests of all Foundations
sub-projects with a single command line as an integral unit. This
tremendously trim the time it takes to build & test Foundations.