WFLoggingImageProvider: fix timeouts in waitForCompletion.
Timeouts were caused by wrong assumptions regarding -runMode:beforeDate:
It does not run until the given date, but runs the runloop once, with
the date used as hard timeout. Thus, the counter of total time spent
was much higher than the actual time, leading to seemingly random
timeouts.

The fix is not to accumulate running time, but to test current time for
timeout. Usage of NSDate is avoided, because it depends on system time
which can change. Instead mach_absolute_time() is used (via
CACurrentMediaTime).

Another potential fix is to use -runUntilDate: which does block
until the date. The drawback is that still manual iterations and timeout
polling is needed, and so spinning the runloop once each time makes the
tests complete a little faster.