LTPresentationViewSpec: fix crash in glDeleteFramebuffers on device.
The crash occurs in "should forward event to delegate" test, when
view's frame buffer is destroyed. The problem seems to be that
the frame buffer as used in that test is backed by texture, and
not renderbuffer. As an optimization, the view discards the frame buffer
after each rendering pass, and this is what we believed causes the later
crash.

As a solution, we create the real instance of the view, back by actual
frame buffer, crated from a drawable provided by LTEAGLView.
This has the added benefit of actually testing the view's response to
bounds change, and gets rid of a hack where unrelated mock of LTEAGLView
was being injected into the view, via cast that exposes an internal
implementation details (the view implementing  LTEAGLViewDelegate
protocol).