LTCGExtensions: add CGRect rounding to pixel boundaries.
Current implementation can only round to integral points. This commit
adds scale parameter to the family of CGRountRect* function, that is
used to convert from point to pixels.

In addition, the existing implementation could not handle non positive
sizes, and CGRoundRectInside was not well defined, as it could return
negative sized rects event for positive size input. This is also fixed.