LTMMTexture: Add glFlush() in order to fix bug of glClientWaitSync.
This commit adds the glFlush() statement before the call to the
glClientWaitSync(APPLE) function performed in LTMMTexture. This is
necessary since there seems to be a bug in the implementation of
glClientWaitSync(APPLE) occurring on devices running iOS 10. The
function seems not to perform the glFlush() equivalent it is required to
perform according to the OpenGL specification.