LTCVPixelBufferExtensions: unify planar and non planar buffers.
Current API for reading & writing pixel buffers is different based of
whether the buffer is non planer (single matrix) or planar (single
matrix per plane, identified by index). In some case, it is desirable to
have unified API.

This commit introduces LTCVPixelBufferImages family of functions, that
pass std::vector<cv::Mat> with matrix per plane for planar buffers and
a single matrix for non-planar buffer.

There are a couple of benefits with this approach:
- Less boilerplate when all planes are needed at once.
- Non-planar buffer becomes somewhat similar to a single plane planar
  buffer.