LTTouchEventView: Handle problematic value of coalescedTouchesForTouch:.
This commit updates the LTTouchEventView to correctly deal with the rare
but existing case that the call to the coalescedTouchesForTouch: method
returns nil even when both the UIEvent on which the method is called and
the UITouch provided as parameter are not nil. The case is specifically
handled by returning an LTTouchEvent constructed from the main touch
instead.