DVNPainter: Initial commit.
Object representing a "painter", i.e. an entity with a "canvas" onto
which it draws with a "brush" along a "stroke". The canvas is
represented by an LTTexture, while the brush is described by a
DVNPipelineConfiguration, and the stroke is incorporated by a sequence
of LTSplineControlPoint objects. For every stroke, a different brush can
be used.