NSValue+LTQuad: Initial commit.
This commit adds NSValue+LTQuad, a category augmenting the NSValue class
with convenience methods for boxing and unboxing of lt::Quad values.
Boxing and unboxing of lt::Quad is desired since a) the usage of the
LTQuad class should be minimized and b) it must be possible to create
stubbable/mockable methods having lt::Quad (or its boxed version) as
parameters.