Bazaar: refresh receipt when purchased product is not found in it.
A few customers had problems after purchasing a subscription: even
though the purchase was successful, the subscription wasn't in the
receipt returned by Validatricks.
According to Apple's In-App Purchase FAQ, it may be the case where the
receipt is not updated yet and needs to be refreshed.
I've added code that refreshes the receipt in case the product that was
just purchased successfully is not found in the receipt received from
Validatricks. After refereshing, another validation with Validatricks
is invoked. In addition, an event is sent whenever this scenario
 occurs, so we'll be able to monitor it.