LTQuadMapping: Add methods using lt::Quad and update docs.
This commit augments the LTQuadMapping file with the same methods as the
existing ones but using lt::Quad. This is desirable since the LTQuad
class is currently restricted to corners provided in clockwise
direction, a restriction which is obstructive when attempting to render
to back-facing quads. In the scope of this PR, the docs of the file are
updated as well. Note that no existing functionality is changed by this
commit.