LTMixerProcessor: Update frontQuad of internal processor when required.
This commit updates the LTMixerProcessor to update the frontQuad of the
internally used processor in the preprocess method, when required. The
change is necessary due to the fact that the frontScaling property of
the processor is now restricted to positive values. Without this change,
calls to the resetToInputModel method can (depending on the order of the
reset keys) result in the creation of an lt::Quad with zero size.