PTNDescriptor: Add RAW image trait.
This commit adds a new trait to PTNDescriptor - RAW image. The trait
specifies that an asset has a RAW image. The trait is implemented only
in PhotoKit assets. API to fetch RAW image for an asset will be added
in a later commit.
This commit also includes a fix to a test case that was causing the
newly added tests to fail because a mocked NSURL object never died due
to a retain cycle.