BZRAllowedProductsProvider: Fix subscription product filtering.
Users that own a subscription are eligible to some sub-set of the
products in the the product list. Each subscription product grants a
different sub-set, what’s common to all these sub-sets is that they
should exclude subscription products. This commit fixes the filtering
of subscription products from the sub-set of products that the user is
eligible for.