LTLogger: register target with minimal log level.
This is an API-breaking change that associates a specific target with
its minimal log level. It will allow registering different targets that
require different minimal log level configuration, and hence will
extract that logic from the target itself to the logger. This benefits
in better separation of concerns and performance, since the logger can
avoid formatting messages that will not be logged by any target.