LTKit: remove NSObject+NSSet.
This code is removed for the following reasons:

1. This category conflicts with `lt_set` in `NSArray+NSSet`.
2. The shortening in creating a set with a single object is augmented
with cognitive overhead and possible bugs that may be created by the
user of this category. For example, it's not immediately clear what are
the results of the following expressions:
```
[@7 lt_set]
[@[@7] lt_set]
[@[@[@7]] lt_set]
[[@[@[@7]] lt_set] lt_set]
```

Therefore, we decided it's better to use `-[NSObject setWithObject:]`
or `[@[object] lt_set]` instead.