DVNBrush: Correctly compute sample locations and improve performance.
This commit updates the DVNBrush vertex and fragment shaders such that
the locations at which the edge avoidance guide texture is to be sampled
can be computed in the correct way. Previously, the sample locations
were based on the rendered quad which yielded wrong results for
transformed quads. In addition to the fix, the performance of the shader
execution is improved by sampling the edge avoidance guide texture in
the vertex shader rather than the fragment shader.