LTContentNavigationManager: Add minZoomScaleFactor property.
Add minZoomScaleFactor property, a multiplicative factor to adjust the
minimum zoom scale for the content. A value of 1 causes the content to
be aspect-fit to the bounds of the view managed by this instance, upon
minimum zoom scale. Values lower than allow the content to be
proportionally smaller than the aspect-fit rectangle, while values
greater than restrict the content to be proportionally larger than the
aspect-fit rectangle. Initial value is 1.