LTMMTexture: wait on GPU read followed by a CPU write.
This fixes a long standing bug in LTMMTexture, which is manifested when
performing a GPU read before a CPU write. Since no synchronization was
performed, a potential out-of-order execution could've happened,
swapping the GPU read with the CPU write, yielding a GPU read of a
buffer with invalid contents.

The solution is to introduce a second sync object that will be pushed to
The pipeline when a GPU read operation is performed. Then upon mapping
the texture for writing (only, reading is safe), the thread will block
until the reading has been done executing.

Due to the nature of the operations that are usually done on textures,
this specific order was rarely performed (since the more common flows
are CPU write followed by a GPU read or GPU write followed by a CPU
read). However existing code that rely on texture CPU mapping should be
tested for performance after this change is committed.