Bazaar Tests: Fix analyzer warnings.
Some of the tests in `BZRStoreSpec` assigned a value to `weakStore`
variable but never used that variable after the assignment. After
reading the tests it seemed like one of the tests was cumbersome and
could be split into two different tests. So this commit split the test
and avoid the use of `weakStore` when not really needed.