DVNBrush.fsh: Fix bugs and optimize code.
The spatial factor used in the edgeAvoidanceFactor method should only be
based on the brush tip texture, analogous to the implementation in
LTEdgeAvoidingBrush.fsh. As part of this commit, a minor optimization of
the fragment shader is performed, by avoiding redundant
premultiplication and the DVNBrushSpec file is added, intended to test
the DVNBrush vertex and fragment shader.