LTAssert: log error only in Debug configurations.
If an assert has failed, the code is already going to crash, and since
the exception contains the logged message there's no need to duplicate
it. In Debug configurations, however, this may be useful, since the
exception message is not always very visible.

Another good reason for removing the logging in non-Debug configurations
is the binary size. Since assertions are very popular across the
codebase, many logging strings are expected to be generated by asserts.
Therefore, removing these logs may also have an impact on the binary
size.