Bazaar: do not cancel request if the signal wasn't disposed of.
Some users experienced crashes in the middle of drawing product list
prices from Apple. This happened when `cancel` of `SKProductsRequest`
was called. Until now, `cancel` was always called after the signal of
the request has completed or erred, which might be causing the problem
since the request was accessed after it should have been released.