FBRAFNetworkingSessionAdapter: Dispatch main queue if not provided any.
If the `sessionManager.completionQueue` is nil, the manager is using the
main queue (by default). However, on the creation of an HTTP task, if
the serialization of the request of the creation of the task fails
(very unlikely) the adapter dispatch queue the failure block on
sessionManager.completionQueue which is nil.