LLSignalTestRecorder: avoid retain cycles if recorder is active.
The LLSignalTestRecorder held the tested signal, while the signal
implicitly held the recorder. This meant that while the subscription is
active, the retain cycle is in place and both of the objects (including
all the objects that were sent on the signal, which are retained by
LLSignalTestRecorder), will be kept alive.

This solves the issue by making sure the test recorder will deallocate
when it's no longer held, assuring that its -dealloc will be called and
that the subscription will be disposed.