AFURLRequestSerialization: Dispatch the right serializer queue on copy.
## Summary

Suggested fix for a crash that occurs when copying an
`AFHTTPRequestSerializer`. The crash is caused by dispatching a block
that modifies the request headers dictionary of the **new** serializer
instance on the wrong queue - the queue of the serializer being copied
instead of the queue of the new serializer. Due to this error it happens
that this block may replace the request headers dictionary (i.e. assign
a brand new dictionary and release the existing one) while
simultaneously another block, dispatched by the initializer of the new
serializer on the correct queue, adds a new header field to the existing
dictionary.

Typically the stack trace looks like this:
Crashed: requestHeaderModificationQueue
Exception: message sent to deallocated instance <address>
```
0  libobjc.A.dylib   0x18ceb01a0 objc_retain + 16
1  AFNetworking      0x1019d3fa0 __55-[AFHTTPRequestSerializer
                         setValue:forHTTPHeaderField:]_block_invoke + 40
2  libdispatch.dylib 0x18d2ea9e0 _dispatch_call_block_and_release + 24
```

The crash occurs with a chance of about `1/10000` and it can be
reproduced by running `[serializer copy]` in a simple `for` loop with
more than `10000` iterations.

## Details

An `AFHTTPRequestSerializer` instance provides an interface for getting
and setting the HTTP headers for all requests serialized by that
instance. The headers are saved in a mutable dictionary. Every access to
that mutable dictionary is done via `requestHeaderModificationQueue`,
which is a concurrent queue, while mutations are performed with a
barrier. This way the queue acts as a Single-Writer-Multiple-Reader lock
that also promises that writes will execute alone and that any
reads/writes that were queued after a certain write will only start
executing after that write completes and get the updated data.

The scenario that leads to the crash:
- `serializerA` is allocated and initialized. A new mutable headers
dictionary and designated queue are created for it - `headersA` and
`queueA`.
- `[serializerA copy]` is invoked which in turn invokes `copyWithZone:`.
- Inside `copyWithZone:` a new serializer is created and initialized -
`serializerB`.
- Within the initializer of `serializerB` a new headers dictionary is
allocated along with a designated queue - `headersB` and `queueB`.
- Also within the initializer `setValue:forHTTPHeaderField:` is called
which queues a block to run  asynchronously on `queueB`.
- After initializing `serializerB` the `copyWithZone:` method continues
to copy `headersA` and overwrite `headersB` with that copy.
*Before the fix* this copying and overwriting is done by dispatching a
block on `queueA` (and not `queueB`).
- Since the header-setting block was queued on `queueB` and the copy
block was queued on `queueA` it is possible that they will will be
scheduled to run simultaneously. If this happens the copy block will
cause the existing headers dictionary to be released while the
header-setting block is still using that dictionary causing the block
to crash.