LTValueObject: switch operands in -isKindOfClass:.
The previous refactoring of this code switched the `first` and `second`
operands of the `-isKindOfClass:` method. This made the code return
`YES` or crash when `-isEqual:` was called for two objects, where the
receiver object inherits from the argument. For example, the following
code returned `YES`:

```
LTValueObject *object = [[LTValueObject alloc] init];
BOOL equal = [object isEqual:[[NSObject alloc] init]];
```

The fix is to switch the two operands.