NSBundle+OnDemandResources: Add try-catch for ODR access.
In order to overcome a random ODR issue when trying to access the ODR
resources:

```
Fatal Exception: NSInternalInconsistencyException
NSBundleResourceRequest could not connect to its helper daemon
```
1. The try-catch in `fbr_beginAccessToResourcesWithTags` signal should
return error if `beginAccessingResourcesWithCompletionHandler` raised an
exception.
2. The try-catch in `fbr_conditionallyBeginAccessToResourcesWithTags`
should log the error and return `nil` if
`conditionallyBeginAccessingResourcesWithCompletionHandler` raised an
exception.

Read more about the issue:
https://openradar.appspot.com/34389684.