AFURLRequestSerialization: Fix static analyzer issues.
Static analyzer raises the following warning:
"Property of mutable type ‘NSMutableURLRequest’ has ‘copy’ attribute;
an immutable object will be stored instead".

This commit changes the properties of `AFStreamingMultipartFormData`
to be `readonly` as they are only assigned value once during
initialization. The copy is replaced with `strong` and assigned values
that require copying are copied manually and correctly on
initialization.