kLTURLValueTransformer: Avoid percent escaping transformed strings.
This solves a bug where serialized URLs were percent escaped twice.
This was due to wrongfully assuming that in the reverse transform the
url's `absoluteString` property removes escaping from the URL's content.
Thus the string `http://foo%20bar` for example was transformed by the
transformer to the URL `http://foo%2520bar` by escaping `%` to `%25`.