LTTouchEventView: change stationary touch forwarding property to bool.
The previous property backed the `preferredFramesPerSecond` property of
`CADisplayLink`, but the documentation of the property was inconsistent
with the documentation of the display link by specifying that a value
of `0` disables the stationary touch forwarding, while `0` for
`preferredFramesPerSecond` means that the default refresh rate should
be used.

Since all uses of this property are binary (either default value or
`0`), changing the property to `BOOL` simplifies the interface while
preserving the required functionality.