LTGLContext: avoid creating EAGLContext on autoreleasepool.
This change is trying to avoid as much as possible to create an
`EAGLContext` on an `autoreleasepool`. Without it, the following code
returns a `nil` `EAGLContext` after a while:

```
dispatch_queue_t queue = dispatch_queue_create("queue",
    DISPATCH_QUEUE_SERIAL);

for (int i = 0; i < 1000; ++i) {
  dispatch_async(queue, ^{
    auto context = [[[Foo alloc] init] foo];
    LTAssert(context);
  });
}
```

Note that until we connect all the GPU resources and their context in
code, it's still possible that an `LTGPUResource` will be deallocated
outside the context.