PTNMediaLibraryAssetManager: extract all items to an array.
Although `MPMediaQuery` claims that the `items` property is `NSArray`,
this is not the case. The object is `MPMediaEntityResultSetArray`, which
is a subclass of `NSArray` that contains a bunch of state, and
apparently points to mutable data that may change when the media library
changes, as can be seen in this crash: http://crashes.to/s/6d83fe5babd.

An attempt to solve this is to do an explicit copy the array to make
sure we're returning an immutable collection that may not change later
on.