LTImageTIFFCompressor: remove tiling from metadata.
TIFF compression with declared tiling in the image metadata leads to
unexpected observed behaviours:
1) When exporting the image to Instagram, the image is presented as
BGRA.
2) When exporting the image to MAC using Image Capture and then entering
the "get info" window of the image file, the preview version of the
image in the window is also presented as BGRA.

The reason for these behaviours is not certainly known: it might be that
the compressor doesn't compress the image correctly with tiling or that
some decoding method (used by Instagram or the preview image in MAC)
doesn't handle the compress tiled image as it should. The current
solution that works in practice is to remove the tiling declaration
entries from the metadata dictionary.