Bazaar: cache the App Store locale.
Until now, when validation occurred before the prices were fetched, the
App Store locale was not included in the information sent to
Validatricks. The App Store locale helps us identify cases where the
user switched the App Store country and thus we decided to cache it once
it was fetched, so that the chance of it not being available is small.