RACSignal+CloudKitRetry: Fix retry to only retry for retryable errors.
It seems that CloudKit appends the CKRetryAfterKey to the userInfo
dictionary for every CKError and not just to errors that are meant to
be retried (eg. even for CKErrorNotAuthenticated).

Instead of just checking the availability of the CKRetryAfterKey in the
userInfo dictionary we validate that the error is indeed retryable.