LTRef: change semantics to shared pointer.
After two years of using `lt::Ref`, we realize that the initial decision
to treat it as an `std::unique_ptr` that manages Core Foundation objects
is suboptimal:

- The underlying reference is always handled by Core Foundation,
  therefore the object is refcounted no matter what, although we treat
  it not as such.
- Using `lt::Ref` caused a lot of confusion across developers as
  sometimes they've had multiple `lt::Ref`s covering the same reference,
  an invariant that should never happen with std::unique_ptr.
- Since the previous implementation was not copy constructible, adding
  a field of `lt::Ref` to a struct prohibited it from being copyable,
  making some APIs cumbersome.

The new implementation adds the ability to copy construct and to assign,
while removing redundant code for supporting arbitrary object types.
Additionally, we introduce the `::retain()` method to support retaining
a reference, to avoid developers from using the low level `CFRetain`
function in their code.

The performance implications of this change is, on the surface, non
existent, because we needed to handle refcounting in the original
solution as well. However, since copy construction and assignment are
now available, developers may use the underlying reference counting
mechanism more than expected. We consider this change to be negligible
considering the upsides of the change.