PTNOceanAssetManager: Prefer the asset with the biggest pixel count.
Ocean descriptor may have few versions with various sizes to choose from
when fetching the asset. As a result, there are scenarios in which there
are few good candidates to choose from. The changes introduce by this
commit make sure that in such cases the asset with the biggest pixel
count will be chosen.