SPXAlertViewModel: Change action prop from RACCommand to LTVoidBlock.
Due to the way alerts are presented currently in Shopix it is not very
usable to have a RACCommand, it is a lot easier to work with a block
instead.

We should consider changing SPXSubscriptionManager interface to return
a signal instead of getting a completion block, and then maybe we
will be able to work with RACCommand. Anyway using LTVoidBlock as the
button action gives even more flexibility and allows as to mimic
RACCommand execution if needed by subscribing to a signal that performs
the desired work inside the action block.