SPXSubscriptionManager: Use `SPXAlertViewControllerProvider`.
Instead of generating the `UIAlertViewController` inside
`SPXSubscriptionManager` it uses `SPXAlertViewModel+ShopixPresets` to
get an `SPXAlertViewModel` it then passes it to the
`SPXAlertViewControllerProvider` and presents the returned alert
controller.

This is another step towards making the subscription manager clean of UI
code, eventually it will only generate view-models, dispatch them on a
designated signal and the view layer will use these models to present
the desired UI (alerts / feedback mail composer).