LTNullability: allow to return a convertible default value.
This allows to return a default value that is convertible to the
original object that was provided. A third template argument on the
return type is not provided to avoid scenarios where the return type
needs to be explicitly stated. This should be redundant since the only
thing that `lt::nn` does is to strip the `_Nullable` decoration from the
input type, and should not be involved in an additional upcast.