BZRStore: insert manually fetched products to product dictionary.
In order to enable purchasing products that were fetched manually, it is
needed to cache them in `productDictionary` as done with the products
fetched using `productsProvider`.
In addition, to preserve the current behavior of `productList`, a
boolean variable was added differentiating between whole product list
fetching and specific products fetching.