SPXProductDescriptorFactory: Fix error out-parameter nullability issues.
1. Verify that the `NSError **error` out-parameter is not `nil` before
assigning a value to it.
2. Change `validatePromotion:` method to not return `nullable NSError *`
instead it returns `BOOL` and gets an `NSError **` and fill in the error
pointer if needed. This is the more common way to pass errors in Cocoa
and this spares the need to create `NSError` inside the method just in
order to pass it to underlying methods that accept it as out-parameter.