PhotoKit+Photons: Improve cloud based traits.
It appears that the undocumented property `isCloudPlaceholder` doesn't
cover all the cases where an asset is stored only in the cloud. Reverse
engineering the method yielded the following:

```
Photos`-[PHAsset isCloudPlaceholder]:
    0x196f945f0 <+0>:  stp    x29, x30, [sp, #-0x10]!
    0x196f945f4 <+4>:  mov    x29, sp
    0x196f945f8 <+8>:  adrp   x8, 103738
    0x196f945fc <+12>: ldr    x1, [x8, #0xe80]
    // [self cloudPlaceholderKind]
    0x196f94600 <+16>: bl     0x1912de530
    0x196f94604 <+20>: orr    x8, x0, #0x1
    0x196f94608 <+24>: cmp    x8, #0x3                  ; =0x3
    0x196f9460c <+28>: cset   w0, eq
    0x196f94610 <+32>: ldp    x29, x30, [sp], #0x10
    0x196f94614 <+36>: ret
```

Which means that `isCloudPlaceholder` isn't the source of truth, but
fetches `cloudPlaceholderKind` (which is backed by an ivar) and
compares it to `3`. After some experimentation it became clear that
a value of `4` also represents an asset that is stored in the cloud, so
the condition changed to rely on the new `cloudPlaceholderKind`
property instead with these two values.